Refactor test : les test dans le dossier form sont des test sur la vue contenant les formulaire -> decplacement dans les test sur les vues
+Ajout d'un test sur le formulaire nouvelle_formation_form (test KO)
