+++ formation_metier/tests/view/test_detail_formation_view.py
