+++ formation_metier/tests/view/test_detail_seance_view.py
