+++ formation_metier/tests/view/test_list_formation_view.py
